Switch simulation state to float64; default parallel kernel to one thread per worker
Inventory, capacity, NN params/activations, and capacity-delta scratch now use
double instead of float across both the Picard (parallel) and sequential
paths, plus the shared benchmark kernels. The .npy loader no longer downcasts
the already-float64 source data.

Also flips the default fulfillment kernel in the Picard loop from the
warp-cooperative variant to the single-thread-per-worker scalar kernel
(bit-identical scan); the warp kernel remains available via WARP_KERNEL=1.

+++ include/fulfillment.h
@@ -8,17 +8,17 @@
 // with NN output as infinitesimal tie-breaker on inventory.
 // This is the default policy from jax-sim-v2.py (_fulfillment_greedy_cost).
 __device__ inline int fulfillment_greedy_cost(
-    const float* nn_params,
+    const double* nn_params,
     const int* nn_layer_sizes,
     int nn_num_layers,
-    const float* inventory_row,    // [n_nodes_plus1] for this product
-    const float* capacity,         // [n_nodes_plus1]
+    const double* inventory_row,  // [n_nodes_plus1] for this product
+    const double* capacity,       // [n_nodes_plus1]
     int quantity,
     const int* node_near_to_far,   // [n_nodes_plus1] sorted node indices
     int n_nodes_plus1,
-    float* scratch_a,              // scratch buffer for NN: max(layer_sizes)
-    float* scratch_b,              // scratch buffer for NN: max(layer_sizes)
-    float* nn_output               // scratch buffer for NN output: layer_sizes[last]
+    double* scratch_a,             // scratch buffer for NN: max(layer_sizes)
+    double* scratch_b,              // scratch buffer for NN: max(layer_sizes)
+    double* nn_output               // scratch buffer for NN output: layer_sizes[last]
 ) {
     int n_nodes = n_nodes_plus1 - 1;  // exclude unfulfill sentinel
 
@@ -27,19 +27,19 @@ __device__ inline int fulfillment_greedy_cost(
                inventory_row, nn_output, scratch_a, scratch_b);
 
     // Compute norm of NN output
-    float norm = 0.0f;
+    double norm = 0.0;
     for (int i = 0; i < n_nodes_plus1; i++) {
         norm += nn_output[i] * nn_output[i];
     }
-    norm = sqrtf(norm + 1e-12f);
+    norm = sqrt(norm + 1e-12);
 
     // For each node in near-to-far order, check feasibility
     // inventory is tie-broken with 1e-7 * output / norm
     for (int rank = 0; rank < n_nodes_plus1; rank++) {
         int node = node_near_to_far[rank];
-        float inv = inventory_row[node] + roundf(1e-7f * nn_output[rank] / norm);
-        float cap = capacity[node];
-        if (cap >= (float)quantity && inv >= (float)quantity) {
+        double inv = inventory_row[node] + round(1e-7 * nn_output[rank] / norm);
+        double cap = capacity[node];
+        if (cap >= (double)quantity && inv >= (double)quantity) {
             return node;
         }
     }
@@ -49,21 +49,21 @@ __device__ inline int fulfillment_greedy_cost(
 
 // Greedy-capacity fulfillment: pick feasible node with maximum remaining capacity.
 __device__ inline int fulfillment_greedy_capacity(
-    const float* inventory_row,    // [n_nodes_plus1] for this product
-    const float* capacity,         // [n_nodes_plus1]
+    const double* inventory_row,  // [n_nodes_plus1] for this product
+    const double* capacity,       // [n_nodes_plus1]
     int quantity,
     const int* node_near_to_far,   // [n_nodes_plus1]
     int n_nodes_plus1
 ) {
     int best_node = -1;
-    float best_cap = -1.0f;
+    double best_cap = -1.0;
 
     // Exclude sentinel (last entry) from argmax
     for (int rank = 0; rank < n_nodes_plus1 - 1; rank++) {
         int node = node_near_to_far[rank];
-        float inv = inventory_row[node];
-        float cap = capacity[node];
-        if (cap >= (float)quantity && inv >= (float)quantity) {
+        double inv = inventory_row[node];
+        double cap = capacity[node];
+        if (cap >= (double)quantity && inv >= (double)quantity) {
             if (cap > best_cap) {
                 best_cap = cap;
                 best_node = node;
@@ -82,18 +82,18 @@ __device__ inline int fulfillment_greedy_capacity(
 // using nn.greedy_cost_prob (Bernoulli draw).
 __device__ inline int fulfillment(
     uint64_t& rng_key,
-    float greedy_cost_prob,
-    const float* nn_params,
+    double greedy_cost_prob,
+    const double* nn_params,
     const int* nn_layer_sizes,
     int nn_num_layers,
-    const float* inventory_row,
-    const float* capacity,
+    const double* inventory_row,
+    const double* capacity,
     int quantity,
     const int* node_near_to_far,
     int n_nodes_plus1,
-    float* scratch_a,
-    float* scratch_b,
-    float* nn_output
+    double* scratch_a,
+    double* scratch_b,
+    double* nn_output
 ) {
     // In jax-sim-v2.py, fulfillment always calls _fulfillment_greedy_cost
     // (the greedy_cost_prob / greedy_capacity branch is in v2's _fulfillment_greedy)
